requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
APScheduler==3.10.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
billiard==3.6.4.0
black==23.3.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.2.35
Django==4.1.7
django-apscheduler==0.6.2
django-celery-beat==2.5.0
django-cors-headers==3.14.0
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
environs==9.5.0
exceptiongroup==1.1.1
frozenlist==1.3.3
gevent==22.10.2
graypy==2.1.0
greenlet==2.0.2
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
kombu==5.3.0b3
magic-filter==1.0.9
MarkupSafe==2.1.2
marshmallow==3.19.0
multidict==6.0.4
mypy-extensions==1.0.0
mysqlclient==2.1.1
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.0
pluggy==1.0.0
prompt-toolkit==3.0.38
pycparser==2.21
PyJWT==2.6.0
pyTelegramBotAPI==4.11.0
pytest==7.3.1
pytest-django==4.5.2
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
redis==4.5.4
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
telebot==0.0.5
tomli==2.0.1
tzdata==2022.7
tzlocal==4.3
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
yarl==1.8.2
zope.event==4.6
zope.interface==6.0
